Moved Chicken parser and renamed stuff
